backend/requirements.txt
fastapi
uvicorn
langchain
langchain-experimental
langchain-groq
langchain-community
pandas
matplotlib
seaborn
python-dotenv
pydantic